Reorganize and DRY up
