+++ src/tuple.js
@@ -1,7 +1,7 @@
 // @flow strict
 
 import { annotate } from './lib/Annotation';
-import { compose, predicate } from './utils';
+import { compose, predicate } from './composition';
 import { err, errValue, isErr, ok, unwrap, value } from './lib/Result';
 import { poja } from './array';
 import type { Decoder } from './types';
